moved printing table of DataRow to table file
