+++ pkg/cmd/worker/shared/view.go
@@ -3,13 +3,12 @@ package shared
 import (
 	"fmt"
 	"github.com/OctopusDeploy/cli/pkg/cmd"
-	"github.com/OctopusDeploy/cli/pkg/cmd/shared"
 	"github.com/OctopusDeploy/cli/pkg/machinescommon"
 	"github.com/OctopusDeploy/cli/pkg/output"
 	"github.com/OctopusDeploy/go-octopusdeploy/v2/pkg/machines"
 )
 
-type ContributeEndpointCallback func(opts *ViewOptions, endpoint machines.IEndpoint) ([]*shared.DataRow, error)
+type ContributeEndpointCallback func(opts *ViewOptions, endpoint machines.IEndpoint) ([]*output.DataRow, error)
 
 type ViewFlags struct {
 	*machinescommon.WebFlags
@@ -41,15 +40,15 @@ func ViewRun(opts *ViewOptions, contributeEndpoint ContributeEndpointCallback, d
 		return err
 	}
 
-	data := []*shared.DataRow{}
+	data := []*output.DataRow{}
 
-	data = append(data, shared.NewDataRow("Name", fmt.Sprintf("%s %s", output.Bold(worker.Name), output.Dimf("(%s)", worker.GetID()))))
-	data = append(data, shared.NewDataRow("Health status", getHealthStatus(worker)))
-	data = append(data, shared.NewDataRow("Current status", worker.StatusSummary))
+	data = append(data, output.NewDataRow("Name", fmt.Sprintf("%s %s", output.Bold(worker.Name), output.Dimf("(%s)", worker.GetID()))))
+	data = append(data, output.NewDataRow("Health status", getHealthStatus(worker)))
+	data = append(data, output.NewDataRow("Current status", worker.StatusSummary))
 
 	workerPoolMap, err := GetWorkerPoolMap(opts)
 	workerPoolNames := resolveValues(worker.WorkerPoolIDs, workerPoolMap)
-	data = append(data, shared.NewDataRow("Worker Pools", output.FormatAsList(workerPoolNames)))
+	data = append(data, output.NewDataRow("Worker Pools", output.FormatAsList(workerPoolNames)))
 
 	if contributeEndpoint != nil {
 		newRows, err := contributeEndpoint(opts, worker.Endpoint)
@@ -61,11 +60,7 @@ func ViewRun(opts *ViewOptions, contributeEndpoint ContributeEndpointCallback, d
 		}
 	}
 
-	t := output.NewTable(opts.Out)
-	for _, row := range data {
-		t.AddRow(row.Name, row.Value)
-	}
-	t.Print()
+	output.PrintRows(data, opts.Out)
 
 	fmt.Fprintf(opts.Out, "\n")
 	machinescommon.DoWebForWorkers(worker, opts.Dependencies, opts.WebFlags, description)
@@ -74,24 +69,24 @@ func ViewRun(opts *ViewOptions, contributeEndpoint ContributeEndpointCallback, d
 	return nil
 }
 
-func ContributeProxy(opts *ViewOptions, proxyID string) ([]*shared.DataRow, error) {
+func ContributeProxy(opts *ViewOptions, proxyID string) ([]*output.DataRow, error) {
 	if proxyID != "" {
 		proxy, err := opts.Client.Proxies.GetById(proxyID)
 		if err != nil {
 			return nil, err
 		}
-		return []*shared.DataRow{shared.NewDataRow("Proxy", proxy.GetName())}, nil
+		return []*output.DataRow{output.NewDataRow("Proxy", proxy.GetName())}, nil
 	}
 
-	return []*shared.DataRow{shared.NewDataRow("Proxy", "None")}, nil
+	return []*output.DataRow{output.NewDataRow("Proxy", "None")}, nil
 }
 
-func ContributeAccount(opts *ViewOptions, accountID string) ([]*shared.DataRow, error) {
+func ContributeAccount(opts *ViewOptions, accountID string) ([]*output.DataRow, error) {
 	account, err := opts.Client.Accounts.GetByID(accountID)
 	if err != nil {
 		return nil, err
 	}
-	data := []*shared.DataRow{shared.NewDataRow("Account", account.GetName())}
+	data := []*output.DataRow{output.NewDataRow("Account", account.GetName())}
 	return data, nil
 }
